requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-dotenv==1.0.0
anthropic==0.5.0
supabase==1.0.4
pydantic==2.4.2
python-multipart==0.0.6
pytest==7.4.3 